requirements.txt
Django==4.2.29
django-rest-framework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
python-dotenv
openai
chromadb
crewai
Pillow
pdfplumber
